RUN-912: Add methods needed for merged plugins
I finally got access to azure so I was able to start testing, and I encountered a problem. With GCP, oracle, and AWS, once I moved the cloud plugins into the main repo, they were all in one place so they shared manager and manager builder files. But with the azure ones, many live in this repo (https://github.com/rundeck-plugins/rundeck-azure-plugin). This is different from the cloud repo, and I was told this one needs to stay on a separate repo because it is open source. However, the builder for all azure plugins uses the files from this repo. When the cloud plugins were separate they had their own, but with me moving them now they will be linked to that one, losing the methods they need. The changes cannot be tested until this separate repo has this PR to add these methods, and the PR is merged to it’s main. But regardless I didn’t change any logic or the actual plugin itself, so the original testing when these plugins were written should still apply once that is sorted out.